Create rnn.asm
Neural Network Implementation

The neural network is implemented using a recurrent neural network (RNN). The RNN has an input layer, a hidden layer, and an output layer. The input layer is the size of the input data, the hidden layer is 50 nodes, and the output layer is the size of the output data.

The RNN is trained using backpropagation. The backpropagation algorithm updates the weights and biases of the neural network to minimize the error between the actual output and the desired output.